Revert "course add page"
