+++ Sootrank/settings.py
@@ -78,11 +78,10 @@ DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',
         'NAME': 'sootrank_db',
-        'USER': 'postgres',
-        'PASSWORD': 'nitin@098',
+        'USER': 'sootrank_user',
+        'PASSWORD': 'YO123#yo',
         'HOST': 'localhost',
         'PORT': '5432',
-        'CONN_MAX_AGE': 0,
     }
 }
 
@@ -132,6 +131,3 @@ STATICFILES_DIRS=[
 # https://docs.djangoproject.com/en/5.2/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
-
-MEDIA_URL = '/media/'
-MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
