--- templates/Registration/add_course_offering.html
@@ -1,188 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Add Course Offering</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-  <style>
-    /* Moving animated techy background */
-    body {
-      font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-      min-height: 100vh;
-      margin: 0;
-      overflow-x: hidden;
-      background: linear-gradient(-45deg, #0b1d51, #1a2a6c, #0f3057, #1e3c72);
-      background-size: 400% 400%;
-      animation: gradientBG 20s ease infinite;
-      color: #ffffff;
-    }
-
-    @keyframes gradientBG {
-      0% { background-position: 0% 50%; }
-      50% { background-position: 100% 50%; }
-      100% { background-position: 0% 50%; }
-    }
-
-    /* Main Card */
-    .card-main {
-      position: relative;
-      z-index: 1;
-      max-width: 800px;
-      margin: 50px auto;
-      padding: 30px;
-      border-radius: 15px;
-      background: rgba(20, 30, 50, 0.95);
-      box-shadow: 0 10px 30px rgba(0,0,0,0.6);
-      backdrop-filter: blur(5px);
-      border: 1px solid rgba(255,255,255,0.1);
-    }
-
-    h2 {
-      font-weight: 600;
-      color: #ffffff;
-    }
-
-    /* Year Containers */
-    .year-container {
-      border-radius: 10px;
-      padding: 15px 20px;
-      margin-bottom: 15px;
-      transition: all 0.3s ease;
-      cursor: pointer;
-    }
-
-    .branch-section {
-      display: none;
-      margin-top: 10px;
-      padding: 10px;
-      background-color: rgba(255,255,255,0.05);
-      border-radius: 8px;
-      animation: fadeIn 0.5s ease;
-    }
-
-    @keyframes fadeIn {
-      from { opacity: 0; transform: translateY(-10px); }
-      to { opacity: 1; transform: translateY(0); }
-    }
-
-    .form-check-input:checked {
-      accent-color: #0d6efd;
-    }
-
-    .btn-primary {
-      background-color: #0d6efd;
-      border-color: #0d6efd;
-      transition: 0.3s;
-    }
-
-    .btn-primary:hover {
-      background-color: #0b5ed7;
-      border-color: #0b5ed7;
-    }
-
-    .year-label {
-      font-weight: 500;
-      color: #ffffff;
-    }
-
-    /* Techy dark shades for each year */
-    .year-1 { background-color: #1f4068; }
-    .year-2 { background-color: #2b5876; }
-    .year-3 { background-color: #3a6ea5; }
-    .year-4 { background-color: #4a85c3; }
-
-    @media (max-width: 576px) {
-      .branch-section {
-        padding: 8px;
-      }
-    }
-
-    /* Input & labels in dark theme */
-    .form-label {
-      color: #cfd8dc;
-    }
-
-    .form-check-label {
-      color: #e0e0e0;
-    }
-
-    select.form-select {
-      background-color: #1e2a5c;
-      color: #ffffff;
-      border: 1px solid #3a6ea5;
-    }
-
-    select.form-select:focus {
-      box-shadow: 0 0 0 0.25rem rgba(13,110,253,.25);
-      border-color: #0d6efd;
-    }
-  </style>
-</head>
-<body>
-
-<div class="container">
-  <div class="card card-main">
-    <h2 class="mb-4 text-center">Add Course Offering</h2>
-
-    <form method="post">
-      {% csrf_token %}
-
-      <!-- Course Code Dropdown -->
-      <div class="mb-4">
-        <label class="form-label">{{ form.course_code.label }}</label>
-        {{ form.course_code }}
-      </div>
-
-      <!-- Year checkboxes in dark techy containers -->
-      <div class="mb-4">
-        <label class="form-label">Select Years & Branches</label>
-        <div>
-          {% for value, label in form.fields.years.choices %}
-            <div class="year-container year-{{ value }}">
-              <div class="form-check mb-2">
-                <input type="checkbox" class="form-check-input year-checkbox" id="year{{ value }}" value="{{ value }}" name="years">
-                <label for="year{{ value }}" class="form-check-label year-label">{{ label }} Year</label>
-              </div>
-
-              <div class="branch-section" id="branches-for-{{ value }}">
-                <strong>Select Branches</strong>
-                <div class="d-flex flex-wrap mt-2">
-                  {% for branch in form.fields.branches.queryset %}
-                    <div class="form-check me-3 mb-2">
-                      <input type="checkbox" class="form-check-input" 
-                             name="branches_{{ value }}" value="{{ branch.id }}" 
-                             id="branch-{{ value }}-{{ branch.id }}">
-                      <label class="form-check-label" 
-                             for="branch-{{ value }}-{{ branch.id }}">{{ branch.name }}</label>
-                    </div>
-                  {% endfor %}
-                </div>
-              </div>
-            </div>
-          {% endfor %}
-        </div>
-      </div>
-
-      <div class="text-center">
-        <button type="submit" class="btn btn-primary btn-lg">Save Offering</button>
-      </div>
-    </form>
-  </div>
-</div>
-
-<script>
-  document.querySelectorAll('.year-checkbox').forEach(cb => {
-    cb.addEventListener('change', function() {
-      let branchSection = document.getElementById('branches-for-' + this.value);
-      if (this.checked) {
-        branchSection.style.display = 'block';
-      } else {
-        branchSection.style.display = 'none';
-        branchSection.querySelectorAll('input[type=checkbox]').forEach(b => b.checked = false);
-      }
-    });
-  });
-</script>
-
-</body>
-</html>
